(PC-41905)[PRO] refactor: separate collective offers from offer templates in CollectiveOfferLayout
