+++ pro/src/pages/CollectiveOffer/CollectiveOffer/CollectiveOfferEdition/__specs__/CollectiveOfferEdition.spec.tsx
@@ -67,7 +67,7 @@ describe('CollectiveOfferEdition', () => {
 
     expect(
       await screen.findByRole('heading', {
-        name: /Modifier l’offre/,
+        name: /Modifier l'offre/,
       })
     ).toBeInTheDocument()
 
@@ -87,7 +87,7 @@ describe('CollectiveOfferEdition', () => {
     })
     expect(
       await screen.findByRole('heading', {
-        name: /Modifier l’offre/,
+        name: /Modifier l'offre/,
       })
     ).toBeInTheDocument()
     expect(screen.getByText('Offre vitrine')).toBeInTheDocument()
