--- pro/src/pages/CollectiveOffer/CollectiveOfferLayout/CollectiveEditionOfferNavigation/CollectiveEditionOfferNavigation.tsx
@@ -1,270 +1,0 @@
-import cn from 'classnames'
-import { useRef, useState } from 'react'
-import { useLocation, useNavigate } from 'react-router'
-import { mutate } from 'swr'
-
-import { apiNew } from '@/apiClient/api'
-import {
-  CollectiveOfferDisplayedStatus,
-  CollectiveOfferTemplateAllowedAction,
-  type GetCollectiveOfferResponseModel,
-  type GetCollectiveOfferTemplateResponseModel,
-} from '@/apiClient/v1/new'
-import { useAnalytics } from '@/app/App/analytics/firebase'
-import {
-  GET_COLLECTIVE_OFFER_QUERY_KEY,
-  GET_COLLECTIVE_OFFER_TEMPLATE_QUERY_KEY,
-} from '@/commons/config/swrQueryKeys'
-import { Events } from '@/commons/core/FirebaseEvents/constants'
-import { isCollectiveOfferTemplate } from '@/commons/core/OfferEducational/types'
-import { computeURLCollectiveOfferId } from '@/commons/core/OfferEducational/utils/computeURLCollectiveOfferId'
-import { createOfferFromTemplate } from '@/commons/core/OfferEducational/utils/createOfferFromTemplate'
-import { duplicateBookableOffer } from '@/commons/core/OfferEducational/utils/duplicateBookableOffer'
-import { useActiveFeature } from '@/commons/hooks/useActiveFeature'
-import { useSnackBar } from '@/commons/hooks/useSnackBar'
-import { isActionAllowedOnCollectiveOffer } from '@/commons/utils/isActionAllowedOnCollectiveOffer'
-import { ArchiveConfirmationModal } from '@/components/ArchiveConfirmationModal/ArchiveConfirmationModal'
-import { ShareLinkDrawer } from '@/components/CollectiveOffer/ShareLinkDrawer/ShareLinkDrawer'
-import { Button } from '@/design-system/Button/Button'
-import {
-  ButtonColor,
-  ButtonSize,
-  ButtonVariant,
-} from '@/design-system/Button/types'
-import fullArchiveIcon from '@/icons/full-archive.svg'
-import fullCopyIcon from '@/icons/full-duplicate.svg'
-import fullPlusIcon from '@/icons/full-plus.svg'
-import fullShowIcon from '@/icons/full-show.svg'
-import type { NavLinkItem } from '@/ui-kit/Tabs/NavLinkItems/NavLinkItems'
-import { Tabs } from '@/ui-kit/Tabs/Tabs'
-
-import { CollectiveOfferStep } from '../CollectiveOfferNavigation/CollectiveCreationOfferNavigation'
-import styles from './CollectiveEditionOfferNavigation.module.scss'
-
-export interface CollectiveEditionOfferNavigationProps {
-  activeStep: CollectiveOfferStep
-  isTemplate: boolean
-  offerId?: number
-  offer?:
-    | GetCollectiveOfferResponseModel
-    | GetCollectiveOfferTemplateResponseModel
-}
-
-export const CollectiveEditionOfferNavigation = ({
-  activeStep,
-  isTemplate = false,
-  offerId = 0,
-  offer,
-}: CollectiveEditionOfferNavigationProps): JSX.Element => {
-  const { logEvent } = useAnalytics()
-  const snackBar = useSnackBar()
-  const navigate = useNavigate()
-  const location = useLocation()
-  const isMarseilleActive = useActiveFeature('ENABLE_MARSEILLE')
-
-  const [isArchiveModalOpen, setIsArchiveModalOpen] = useState(false)
-
-  const id = computeURLCollectiveOfferId(offerId, Boolean(isTemplate))
-
-  const archiveButtonRef = useRef<HTMLButtonElement>(null)
-  const adagePreviewButtonRef = useRef<HTMLAnchorElement>(null)
-
-  const archiveOffer = async () => {
-    if (!offerId) {
-      snackBar.error('L’identifiant de l’offre n’est pas valide.')
-      return
-    }
-    try {
-      if (isTemplate) {
-        await apiNew.patchCollectiveOffersTemplateArchive({
-          body: { ids: [offerId] },
-        })
-        await mutate([GET_COLLECTIVE_OFFER_TEMPLATE_QUERY_KEY, offerId])
-      } else {
-        await apiNew.patchCollectiveOffersArchive({ body: { ids: [offerId] } })
-        await mutate([GET_COLLECTIVE_OFFER_QUERY_KEY, offerId])
-      }
-
-      setIsArchiveModalOpen(false)
-
-      snackBar.success('Une offre a bien été archivée')
-    } catch {
-      snackBar.error('Une erreur est survenue lors de l’archivage de l’offre')
-    }
-  }
-
-  const canEditOffer =
-    !isTemplate &&
-    offer?.displayedStatus !== CollectiveOfferDisplayedStatus.ARCHIVED &&
-    location.pathname.includes('edition')
-
-  const canPreviewOffer = () => {
-    return (
-      isTemplate &&
-      offer?.displayedStatus !== CollectiveOfferDisplayedStatus.ARCHIVED
-    )
-  }
-
-  const canArchiveOffer = () => {
-    if (!offer) {
-      return false
-    }
-
-    return (
-      isTemplate &&
-      isActionAllowedOnCollectiveOffer(
-        offer,
-        CollectiveOfferTemplateAllowedAction.CAN_ARCHIVE
-      )
-    )
-  }
-
-  const canDuplicateOffer = () => {
-    if (!offer) {
-      return false
-    }
-
-    return (
-      isTemplate &&
-      isActionAllowedOnCollectiveOffer(
-        offer,
-        CollectiveOfferTemplateAllowedAction.CAN_DUPLICATE
-      )
-    )
-  }
-
-  const canCreateBookableOffer =
-    offer &&
-    isActionAllowedOnCollectiveOffer(
-      offer,
-      CollectiveOfferTemplateAllowedAction.CAN_CREATE_BOOKABLE_OFFER
-    )
-
-  const canShareOffer =
-    offer &&
-    isActionAllowedOnCollectiveOffer(
-      offer,
-      CollectiveOfferTemplateAllowedAction.CAN_SHARE
-    )
-
-  const tabs: NavLinkItem<string>[] = [
-    {
-      key: CollectiveOfferStep.DETAILS,
-      label: 'Détails de l’offre',
-      url: `/offre/${offerId}/collectif/edition`,
-    },
-    {
-      key: CollectiveOfferStep.STOCKS,
-      label: 'Dates et prix',
-      url: `/offre/${offerId}/collectif/stocks/edition`,
-    },
-    {
-      key: CollectiveOfferStep.INSTITUTION,
-      label: 'Établissement et enseignant',
-      url: `/offre/${offerId}/collectif/etablissement/edition`,
-    },
-  ]
-
-  return (
-    <>
-      <div className={styles['actions-container']}>
-        {canPreviewOffer() && (
-          <Button
-            as="a"
-            variant={ButtonVariant.SECONDARY}
-            color={ButtonColor.NEUTRAL}
-            size={ButtonSize.SMALL}
-            to={`/offre/${id}/collectif${isTemplate ? '/vitrine' : ''}/apercu`}
-            icon={fullShowIcon}
-            ref={adagePreviewButtonRef}
-            label="Aperçu dans ADAGE"
-          />
-        )}
-
-        {canArchiveOffer() && (
-          <Button
-            onClick={() => setIsArchiveModalOpen(true)}
-            icon={fullArchiveIcon}
-            variant={ButtonVariant.SECONDARY}
-            color={ButtonColor.DANGER}
-            size={ButtonSize.SMALL}
-            ref={archiveButtonRef}
-            label="Archiver"
-          />
-        )}
-
-        {canDuplicateOffer() && (
-          <Button
-            variant={ButtonVariant.SECONDARY}
-            color={ButtonColor.NEUTRAL}
-            icon={fullCopyIcon}
-            size={ButtonSize.SMALL}
-            onClick={async () => {
-              logEvent(Events.CLICKED_DUPLICATE_BOOKABLE_OFFER, {
-                offerId,
-                offerStatus: offer?.displayedStatus,
-                offerType: 'collective',
-              })
-              await duplicateBookableOffer(navigate, snackBar, offerId)
-            }}
-            label="Dupliquer"
-          />
-        )}
-
-        {canCreateBookableOffer && (
-          <Button
-            variant={ButtonVariant.SECONDARY}
-            color={ButtonColor.NEUTRAL}
-            icon={fullPlusIcon}
-            size={ButtonSize.SMALL}
-            onClick={() => {
-              if (isTemplate) {
-                logEvent(Events.CLICKED_DUPLICATE_TEMPLATE_OFFER, {
-                  offerId,
-                  offerType: 'collective',
-                  offerStatus: offer.displayedStatus,
-                })
-                // eslint-disable-next-line @typescript-eslint/no-floating-promises
-                createOfferFromTemplate(
-                  navigate,
-                  snackBar,
-                  offerId,
-                  undefined,
-                  isMarseilleActive
-                )
-              }
-            }}
-            label="Créer une offre réservable"
-          />
-        )}
-        {isCollectiveOfferTemplate(offer) && canShareOffer && (
-          <ShareLinkDrawer offerId={offer.id} />
-        )}
-      </div>
-      {canEditOffer && (
-        <Tabs
-          type="links"
-          items={tabs}
-          selectedKey={activeStep}
-          navLabel="Sous menu - offre collective"
-          className={cn(styles['tabs'], {
-            [styles['tabs-active']]: [
-              CollectiveOfferStep.DETAILS,
-              CollectiveOfferStep.STOCKS,
-              CollectiveOfferStep.INSTITUTION,
-            ].includes(activeStep),
-          })}
-        />
-      )}
-      <ArchiveConfirmationModal
-        onDismiss={() => setIsArchiveModalOpen(false)}
-        onValidate={archiveOffer}
-        offer={offer}
-        isDialogOpen={isArchiveModalOpen}
-        refToFocusOnClose={
-          archiveButtonRef.current ? archiveButtonRef : adagePreviewButtonRef
-        }
-      />
-    </>
-  )
-}
