+++ pro/src/pages/CollectiveOffer/CollectiveOfferLayout/CollectiveOfferLayout.module.scss
@@ -1,19 +1,12 @@
 @use "styles/mixins/_size.scss" as size;
 @use "styles/mixins/_rem.scss" as rem;
 
-/* check-unused-css-disable-next-line */
-.eac-layout {
-  &-heading-wrapper {
-    display: flex;
-    flex-direction: column;
-    padding-top: rem.torem(56px);
+.eac-layout-heading-wrapper {
+  display: flex;
+  flex-direction: column;
+  padding-top: rem.torem(56px);
 
-    @media (min-width: size.$tablet) {
-      padding-top: 0;
-    }
-  }
-
-  &-navigation {
-    margin-bottom: var(--size-spacing-xxl);
+  @media (min-width: size.$tablet) {
+    padding-top: 0;
   }
 }
