+++ pro/src/pages/CollectiveOffer/CollectiveOfferLayout/CollectiveOfferTemplateNavigation/CollectiveOfferTemplateNavigation.module.scss
@@ -6,11 +6,6 @@
   margin-bottom: var(--size-spacing-xxl);
 }
 
-.tabs {
-  margin-top: var(--size-spacing-xl);
-  margin-bottom: unset;
-
-  &-active {
-    margin-bottom: var(--size-spacing-xl);
-  }
+.eac-stepper {
+  margin-bottom: var(--size-spacing-xxl);
 }
